# src/swagger_files/QuickCode.Demo.SmsManagerModule.Api-Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickCode - Demo - SmsManagerModule Api (1.0.0.0)",
    "termsOfService": "https://www.quickcode.net",
    "contact": {
      "name": "Üzeyir Apaydın",
      "url": "https://www.quickcode.net"
    },
    "license": {
      "name": "QuickCode.Net License",
      "url": "https://www.quickcode.net"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/sms-manager-module/otp-types/count": {
      "get": {
        "tags": [
          "OtpTypes"
        ],
        "operationId": "api-sms-manager-module-otp-types-count-get",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/sms-manager-module/otp-types/{id}": {
      "get": {
        "tags": [
          "OtpTypes"
        ],
        "operationId": "api-sms-manager-module-otp-types--id--get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpTypesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "OtpTypes"
        ],
        "operationId": "api-sms-manager-module-otp-types--id--put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpTypesDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpTypesDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OtpTypesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "OtpTypes"
        ],
        "operationId": "api-sms-manager-module-otp-types--id--delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/sms-manager-module/otp-types/{otpTypesId}/otp-messages": {
      "get": {
        "tags": [
          "OtpTypes"
        ],
        "operationId": "api-sms-manager-module-otp-types--otpTypesId--otp-messages-get",
        "parameters": [
          {
            "name": "otpTypesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OtpTypesOtpMessagesRestResponseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/sms-manager-module/otp-types/{otpTypesId}/otp-messages/{otpMessagesId}": {
      "get": {
        "tags": [
          "OtpTypes"
        ],
        "operationId": "api-sms-manager-module-otp-types--otpTypesId--otp-messages--otpMessagesId--get",
        "parameters": [
          {
            "name": "otpTypesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "otpMessagesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpTypesOtpMessagesKeyRestResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OtpTypesDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "template": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OtpTypesOtpMessagesKeyRestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "smsSenderId": {
            "type": "integer",
            "format": "int32"
          },
          "otpTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "gsmNumber": {
            "type": "string",
            "nullable": true
          },
          "otpCode": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "expireSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "messageDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "messageSid": {
            "type": "string",
            "nullable": true
          },
          "dailyCounter": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "OtpTypesOtpMessagesRestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "smsSenderId": {
            "type": "integer",
            "format": "int32"
          },
          "otpTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "gsmNumber": {
            "type": "string",
            "nullable": true
          },
          "otpCode": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "expireSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "messageDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "messageSid": {
            "type": "string",
            "nullable": true
          },
          "dailyCounter": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Please enter a valid Api Key",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}